Use new integer types also in 3rd party code

+++ vendorsrc/martinkahr/RemoteControlWrapper/HIDRemoteControlDevice.h
@@ -44,7 +44,7 @@
 	BOOL openInExclusiveMode;
 	BOOL processesBacklog;	
 	
-	int supportedButtonEvents;
+	NSInteger supportedButtonEvents;
 }
 
 // When your application needs to much time on the main thread when processing an event other events
